Make sharding plan explicit in DLRM example
In previous code, DistributedModelParallel was responsible for creating
sharding plans for DLRM. It relied on hard-coded constants to create the
topology for planning (e.g. batch size=512, HBM_CAP=32GB and so on).
This was problematic because it did not reflect the true system
topology. Also, testing different constraints for sharding types and
compute kernels were inconvenient in the previous code.
Thus, we explicitly created sharding plans for DLRM before DMP and
provided some simple options to make life easier.