Fix change_lr logic. Fix training/val batch iter order so datasets don't mix. (#272)
Summary:
Pull Request resolved: https://github.com/facebookresearch/dlrm/pull/272

Line 431 is the fix to change_lr logic.

Lines 440-448, 457 are to fix training/validation batch iteration order and prevent dataset mixing.   This is an alternative approach to D39597993 + D39597988. But here priority is given to minimal code changes to make reviewing quick and easy.     I think this is the way to go.   It would be nice, however, to later see if the code doesn't need to be so coupled.  For example, the intra-epoch validation function (_evaluate) requires a training set iterator to chain two training batches to the validation set. It doesn't seem ideal, but the code works and maybe that's good enough.

Reviewed By: narayanan2004

Differential Revision: D39651826

fbshipit-source-id: bc25a08670b1ba06d31a2bb78fc4749bae615ac2